requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
gunicorn==22.0.0
requests==2.32.3
python-telegram-bot==21.4
Pillow==10.4.0
openpyxl==3.1.5
reportlab==4.2.2
arabic-reshaper==3.0.0
python-bidi==0.4.2
